🎉 WIP create separate TagsIndexPage
