+++ adminSiteClient/AdminApp.tsx
@@ -15,7 +15,7 @@ import { SourceEditPage } from "./SourceEditPage.js"
 import { RedirectsIndexPage } from "./RedirectsIndexPage.js"
 import SiteRedirectsIndexPage from "./SiteRedirectsIndexPage"
 import { TagEditPage } from "./TagEditPage.js"
-import { TagsIndexPage } from "./TagsIndexPage.js"
+import { TagGraphPage } from "./TagGraphPage.js"
 import { PostsIndexPage } from "./PostsIndexPage.js"
 import { TestIndexPage } from "./TestIndexPage.js"
 import { NotFoundPage } from "./NotFoundPage.js"
@@ -42,6 +42,7 @@ import { BulkGrapherConfigEditorPage } from "./BulkGrapherConfigEditor.js"
 import { GdocsIndexPage, GdocsMatchProps } from "./GdocsIndexPage.js"
 import { GdocsPreviewPage } from "./GdocsPreviewPage.js"
 import { GdocsStoreProvider } from "./GdocsStore.js"
+import { TagsIndexPage } from "./TagsIndexPage.js"
 
 @observer
 class AdminErrorMessage extends React.Component<{ admin: Admin }> {
@@ -262,6 +263,11 @@ export class AdminApp extends React.Component<{
                                         tagId={parseInt(match.params.tagId)}
                                     />
                                 )}
+                            />
+                            <Route
+                                exact
+                                path="/tag-graph"
+                                component={TagGraphPage}
                             />
                             <Route
                                 exact
