+++ adminSiteClient/AdminSidebar.tsx
@@ -8,7 +8,6 @@ import {
     faSkullCrossbones,
     faPen,
     faDatabase,
-    faGlobe,
     faTag,
     faUser,
     faArrowRight,
@@ -19,6 +18,7 @@ import {
     faCodeBranch,
     faDownload,
     faHatWizard,
+    faSitemap,
 } from "@fortawesome/free-solid-svg-icons"
 
 import { ETL_WIZARD_URL } from "../settings/clientSettings.js"
@@ -88,6 +88,11 @@ export const AdminSidebar = (): JSX.Element => (
             <li>
                 <Link to="/tags">
                     <FontAwesomeIcon icon={faTag} /> Tags
+                </Link>
+            </li>
+            <li>
+                <Link to="/tag-graph">
+                    <FontAwesomeIcon icon={faSitemap} /> Tag Graph
                 </Link>
             </li>
             <li>
