+++ adminSiteClient/TagGraphPage.scss
@@ -1,4 +1,4 @@
-.TagsIndexPage {
+.TagGraphPage {
     .page-header {
         display: flex;
         justify-content: space-between;
