+++ adminSiteClient/admin.scss
@@ -12,7 +12,7 @@
 @import "codemirror/lib/codemirror.css";
 
 @import "./GdocsIndexPage.scss";
-@import "./TagsIndexPage.scss";
+@import "./TagGraphPage.scss";
 
 html {
     font-size: 14px;
